+++ sourse/Модуль IV Урок 1.md
@@ -1,4 +1,5 @@
-Урок 4. Атака на хранилища данных
+Урок 1. Атака на хранилища данных
+Этот урок открывает четвертый модуль и посвящен атакам на хранилища данных. На практике уязвимости часто спрятаны в глубине приложения – в том месте, где код взаимодействует с хранилищем данных. Глава 9 «Атака на хранилища данных» из книги The Web Application Hacker's Handbook описывает типичные векторы атак на SQL‑, NoSQL‑, XPath‑ и LDAP‑базы. В этом уроке мы обобщим эти идеи, рассмотрим типовые атаки и предложим практические задания на платформе PortSwigger Web Security Academy.ок 4. Атака на хранилища данных
 Этот урок завершает третий модуль и выходит за рамки аутентификации и управления сессиями. На практике уязвимости часто спрятаны в глубине приложения – в том месте, где код взаимодействует с хранилищем данных. Глава 9 «Атака на хранилища данных» из книги The Web Application Hacker’s Handbook описывает типичные векторы атак на SQL‑, NoSQL‑, XPath‑ и LDAP‑базы. В этом уроке мы обобщим эти идеи, рассмотрим типовые атаки и предложим практические задания на платформе PortSwigger Web Security Academy.
 1. Почему хранилище данных – ключевая цель
 Почти все веб‑приложения используют хранилища данных для управления информацией: они хранят учетные записи пользователей, права доступа, настройки конфигурации и другие элементы логики приложения. Современные СУБД – это не просто пассивные контейнеры, они содержат сложную бизнес‑логику и доверяют запросам, поступающим от приложения. Если злоумышленник сможет вмешаться во взаимодействие приложения с хранилищем, он сможет извлечь или изменить конфиденциальные данные и обойти контроль доступа【20†L20-L32】. Поэтому цель атакующего – найти способ превратить пользовательский ввод в часть исполняемого кода на языке запросов.
