+++ src/constants/navigation.ts
@@ -78,15 +78,21 @@ export const navigationLinks: NavLink[] = [
                 href: '/guidelines/module-3/lesson-3',
                 label: 'Урок 3: Атаки на Контроль Доступа',
                 icon: Fingerprint 
-            },
-            {
-                href: '/guidelines/module-3/lesson-4',
-                label: 'Урок 4: Атака на хранилища данных',
-                icon: DatabaseZap
             }
         ]
       },
-      { href: '/guidelines/module-4', label: 'Модуль IV: Серверные Уязвимости', icon: ServerIcon },
+      { 
+        href: '/guidelines/module-4', 
+        label: 'Модуль IV: Серверные Уязвимости', 
+        icon: ServerIcon,
+        children: [
+          {
+            href: '/guidelines/module-4/lesson-1',
+            label: 'Урок 1: Атака на хранилища данных',
+            icon: DatabaseZap
+          }
+        ]
+      },
     ],
   },
   {
